snapm: move Plugin into snapm.manager.plugins
Move the Plugin class into snapm.manager.plugins where it logically
belongs. This simplifies the interface that plugin authors need to be
aware of and avoids the need for plugins to import anything from
snapm.manager.

Resolves: #194, #191
