+++ snapm/manager/plugins/stratis.py
@@ -33,7 +33,6 @@ from snapm import (
     SnapStatus,
     Snapshot,
 )
-from snapm.manager import Plugin
 from snapm.manager.plugins import (
     DEV_PREFIX,
     DMSETUP_CMD,
@@ -41,6 +40,7 @@ from snapm.manager.plugins import (
     DMSETUP_NO_HEADINGS,
     DMSETUP_COLUMNS,
     DMSETUP_FIELDS_UUID,
+    Plugin,
     parse_snapshot_name,
     device_from_mount_point,
     mount_point_space_used,
